requirements.txt
email-validator==1.2.1
Flask>=2.2.3
Flask-WTF>=1.1.1
Flask-Login>=0.6.2
WTForms>=3.0.1
Werkzeug>=2.2.3
mysql-connector-python>=8.0.33
flask-paginate>=2021.1.1
